Remove FeatureFlag.searchImprovements
The flag defaulted to enabled, so the combined-search UI is now the only
code path. Deletes the legacy search results UI (old SearchResultsView,
SearchResultsListView, PodcastsCarouselView, RoundedSubscribeButtonView),
renames NewSearchResultsView to SearchResultsView, extracts the still-used
DisplayMode enum into SearchDisplayMode, and strips the dead episode
search path from SearchResultsModel and the Server module.
