+++ Pocket Casts TV App/UI/Search/SearchViewModel.swift
@@ -15,7 +15,7 @@ enum SearchScope: CaseIterable, Equatable {
         }
     }
 
-    /// Matches the iOS `SearchResultsListView.DisplayMode` analytics values.
+    /// Matches the iOS `SearchDisplayMode` analytics values.
     var analyticsDescription: String {
         switch self {
         case .podcasts:
