+++ podcasts/New Search/Views/Results/PodcastFolderSearchResult+Navigation.swift
@@ -1,0 +1,13 @@
+import PocketCastsDataModel
+import PocketCastsServer
+
+extension PodcastFolderSearchResult {
+    func navigateTo() {
+        switch kind {
+        case .folder:
+            NavigationManager.sharedManager.navigateTo(NavigationManager.folderPageKey, data: [NavigationManager.folderKey: DataManager.sharedManager.findFolder(uuid: uuid) as Any])
+        case .podcast:
+            NavigationManager.sharedManager.navigateTo(NavigationManager.podcastPageKey, data: [NavigationManager.podcastKey: self])
+        }
+    }
+}
