+++ podcasts/New Search/Views/Results/SubscribeButtonView.swift
@@ -3,35 +3,6 @@ import PocketCastsDataModel
 import PocketCastsServer
 import PocketCastsUtils
 
-struct RoundedSubscribeButtonView: View {
-    @ObservedObject var model: SubscribeButtonModel
-
-    init(podcastUuid: String, source: AnalyticsSource, onSubscribe: (() -> Void)? = nil) {
-        self.model = SubscribeButtonModel(podcastUuid: podcastUuid, source: source, subscribeBlock: onSubscribe)
-    }
-
-    var body: some View {
-        Button(action: {
-            if !model.isSubscribed {
-                withAnimation {
-                    model.isSubscribed = true
-                    model.subscribe()
-                }
-            }
-        }) {
-            if model.isSubscribed {
-                Image("discover_subscribed_dark")
-            } else {
-                Image("discover_subscribe_dark")
-            }
-        }
-        .buttonStyle(RoundedSubscribeButtonStyle())
-        .onAppear {
-            model.checkSubscriptionStatus()
-        }
-    }
-}
-
 struct SubscribeButtonView: View {
     @EnvironmentObject var theme: Theme
 
@@ -91,17 +62,6 @@ class SubscribeButtonModel: ObservableObject {
 
     func checkSubscriptionStatus() {
         isSubscribed = DataManager.sharedManager.findPodcast(uuid: podcastUuid) != nil
-    }
-}
-
-private struct RoundedSubscribeButtonStyle: ButtonStyle {
-    func makeBody(configuration: Configuration) -> some View {
-        configuration.label
-        .background(ThemeColor.veil().color)
-        .foregroundColor(ThemeColor.contrast01().color)
-        .cornerRadius(30)
-        .padding([.trailing, .bottom], 6)
-        .applyButtonEffect(isPressed: configuration.isPressed, scaleEffectNumber: 0.8)
     }
 }
