+++ podcasts/New Search/Views/SearchView.swift
@@ -1,5 +1,4 @@
 import SwiftUI
-import PocketCastsUtils
 
 class SearchVisibilityModel: ObservableObject {
     @Published var isSearching = false
@@ -21,11 +20,7 @@ struct SearchView: View {
     @ViewBuilder
     private var searchView: some View {
         if displaySearch.isSearching {
-            if FeatureFlag.searchImprovements.enabled {
-                NewSearchResultsView()
-            } else {
-                SearchResultsView()
-            }
+            SearchResultsView()
         } else {
             SearchHistoryView()
         }
